+++ DummyClient/MovementInterpolator.h
@@ -2,7 +2,7 @@
 
 struct Snapshot
 {
-	float timestamp = 0.f;
+	double timestamp = 0.f;
 	Vec2 position = {};
 	Vec2 velocity = {};
 };
@@ -16,8 +16,8 @@ public:
 	void Update();
 	void Render();
 
-	void SetInterpolationDelay(float delay) { _interpolationDelay = delay; }
-	void SetExtrapolationLimit(float maxDelta) { _extrapolationLimit = maxDelta; }
+	void SetInterpolationDelay(double delay) { _interpolationDelay = delay; }
+	void SetExtrapolationLimit(double maxDelta) { _extrapolationLimit = maxDelta; }
 
 	Vec2& GetLastRenderedPosition() { return _lastRenderedPosition; }
 
@@ -25,21 +25,21 @@ public:
 	void SetBasedOnServerRate();
 
 private:
-	Vec2 Interpolate(float renderTime);
-	Vec2 Extrapolate(float currentTime);
-	bool CanInterpolate(float renderTime);
+	Vec2 Interpolate(double renderTime);
+	Vec2 Extrapolate(double currentTime);
+	bool CanInterpolate(double renderTime);
 
 //	void SetBasedOnServerRate();
 
-	Vec2 Lerp(Vec2& a, Vec2& b, float& t);
-	Vec2 CatmullRom(const Vec2& a, const Vec2& b, const Vec2& c, const Vec2& d, const float& t);
+	Vec2 Lerp(Vec2& a, Vec2& b, double& t);
+	Vec2 CatmullRom(const Vec2& a, const Vec2& b, const Vec2& c, const Vec2& d, const double& t);
 
 private:
 	USE_LOCK
 
 	Deque<Snapshot> _buffer;
-	float _interpolationDelay = 0.f;
-	float _extrapolationLimit = 0.f;
+	double _interpolationDelay = 0.f;
+	double _extrapolationLimit = 0.f;
 
 	Vec2 _lastRenderedPosition = {};
 };
